// src/modules/utils/books-stock/eunm/category.enum.ts
enum CategoryBooksStockEnum {
  ALL = 'All',
  ACTION = 'Action',
  ADVENTURE = 'Adventure',
  BUSINESS = 'Business',
  COMEDY = 'Comedy',
  CRAFTS = 'Crafts',
  CRIME = 'Crime',
  DRAMA = 'Drama',
  GUIDE = 'Guide',
  HEALING = 'Healing',
  HUMOR = 'Humor',
  JOURNAL = 'Journal',
  MUSIC = 'Music',
  ROMANTIC = 'Romantic',
  SPORTS = 'Sports',
  TRAVEL = 'Travel',
}

export default CategoryBooksStockEnum;